+++ src/StateSmith/Runner/FinishedWithFailureException.cs
@@ -1,0 +1,10 @@
+#nullable enable
+
+using System;
+
+namespace StateSmith.Runner;
+
+public class FinishedWithFailureException : Exception
+{
+    public FinishedWithFailureException() : base() { }
+}